Add prerequisite and troubleshooting in READ.md
Add environment setup and Git setup for prerequisite. In order to make
other users understand the processes of this tutorial without teacher'
s guides, add some setup tutorials before ex0 started.

Add "hub: command not found" issue for troubleshooting. When in class,
some students had this problem and spent lots of time fixing it. So I
add this after the prerequisite to solve the problem.